Rearrange packages
