+++ src/main/java/com/algaworks/algafood/domain/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.algaworks.algafood.repository;
+package com.algaworks.algafood.domain.repository;
 
-import com.algaworks.algafood.model.Category;
+import com.algaworks.algafood.domain.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
