+++ src/main/java/com/algaworks/algafood/domain/repository/CityRepository.java
@@ -1,6 +1,6 @@
-package com.algaworks.algafood.repository;
+package com.algaworks.algafood.domain.repository;
 
-import com.algaworks.algafood.model.City;
+import com.algaworks.algafood.domain.model.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CityRepository extends JpaRepository<City, Long> {
