+++ src/main/java/com/algaworks/algafood/domain/repository/PaymentMethodRepository.java
@@ -1,6 +1,6 @@
-package com.algaworks.algafood.repository;
+package com.algaworks.algafood.domain.repository;
 
-import com.algaworks.algafood.model.PaymentMethod;
+import com.algaworks.algafood.domain.model.PaymentMethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Long> {
