+++ src/main/java/com/algaworks/algafood/domain/repository/PermissionRepository.java
@@ -1,6 +1,6 @@
-package com.algaworks.algafood.repository;
+package com.algaworks.algafood.domain.repository;
 
-import com.algaworks.algafood.model.Permission;
+import com.algaworks.algafood.domain.model.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PermissionRepository extends JpaRepository<Permission, Long> {
