+++ src/main/java/com/algaworks/algafood/domain/repository/StateRepository.java
@@ -1,6 +1,6 @@
-package com.algaworks.algafood.repository;
+package com.algaworks.algafood.domain.repository;
 
-import com.algaworks.algafood.model.State;
+import com.algaworks.algafood.domain.model.State;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StateRepository extends JpaRepository<State, Long> {
